Finished tasks

+++ back-end/src/controller/roll-controller.ts
@@ -29,20 +29,29 @@ export class RollController {
   async updateRoll(request: Request, response: Response, next: NextFunction) {
     const { body: params } = request
 
-    this.rollRepository.findOne(params.id).then((roll) => {
-      const updateRollInput: UpdateRollInput = {
-        id: params.id,
-        name: params.name,
-        completed_at: params.completed_at,
-      }
-      roll.prepareToUpdate(updateRollInput)
-      return this.rollRepository.save(roll)
-    })
+    const rollToUpdate = await this.rollRepository.findOne(params.id)
+
+    const updateRollInput: UpdateRollInput = {
+      id: params.id,
+      name: params.name,
+      completed_at: params.completed_at,
+    }
+
+    if (rollToUpdate !== undefined) {
+      rollToUpdate.prepareToUpdate(updateRollInput)
+      return this.rollRepository.save(rollToUpdate)
+    } else {
+      response.status(404).send("Roll not found")
+    }
   }
 
   async removeRoll(request: Request, response: Response, next: NextFunction) {
     let rollToRemove = await this.rollRepository.findOne(request.params.id)
-    await this.rollRepository.remove(rollToRemove)
+    if (rollToRemove !== undefined) {
+      return this.rollRepository.remove(rollToRemove)
+    } else {
+      response.status(404).send("Roll not found")
+    }
   }
 
   async getRoll(request: Request, response: Response, next: NextFunction) {
